requirements.txt
anthropic
google-auth
google-auth-httplib2
google-api-python-client
